# src/main/resources/swagger.yaml
swagger: '2.0'
info:
  description: DataStation API
  version: v0.01
  title: DataStation API
  termsOfService: Terms of service
  contact:
    name: N.A
    url: N.A
    email: N.A
  license:
    name: License of API
    url: API license URL
host: '192.168.210.131:7578'
basePath: /
...
paths:
...
  /news/count/spss/download:
    get:
      tags:
        - news-controller
      summary: getDocumentCount
      description: "검색어가 포함된 뉴스기사의 월별 빈도수 집계 CSV 다운로드(SPSS)"
      operationId: getDocumentCountUsingGET
      produces:
        - '*/*'
      parameters:
        - name: from
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20180101'
        - name: query
          in: query
          description: query
          required: true
          type: string
        - name: to
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20180211'
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT:
            - global
      deprecated: false
  /news/entity:
    get:
      tags:
        - news-controller
      summary: getEntityByName
      description: 검색어가 포함된 뉴스기사의 개체명 항목 
      operationId: getEntityByNameUsingGET
      produces:
        - '*/*'
      parameters:
        - name: entity
          in: query
          description: "\"organizationNamedEntity\", \"locationNamedEntity\", \"etcNamedEntity\", \"totalNamedEntity\""
          required: true
          type: string
        - name: from
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20170101'
        - name: page
          in: query
          description: "from: (page-1)*size, size: size"
          required: true
          type: integer
          format: int32
        - name: query
          in: query
          description: query
          required: true
          type: string
        - name: size
          in: query
          description: "from: (page-1)*size, size: size"
          required: true
          type: integer
          format: int32
        - name: sort
          in: query
          description: "\"entity.asc\", \"entity.desc\", \"date.asc\", \"date.desc\"; default: date.desc,entity.desc"
          required: false
          type: array
          items:
            type: string
          collectionFormat: multi
          default: 'date.desc,entity.desc'
        - name: to
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20190101'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/NewsNamedEntityList'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT:
            - global
      deprecated: false
  /news/entity/download:
    get:
      tags:
        - news-controller
      summary: downloadEntityByName
      description: 검색어가 포함된 뉴스기사의 개체명 항목 다운로드 
      operationId: downloadEntityByNameUsingGET
      produces:
        - '*/*'
      parameters:
        - name: entity
          in: query
          description: "\"organizationNamedEntity\", \"locationNamedEntity\", \"etcNamedEntity\", \"totalNamedEntity\""
          required: true
          type: string
        - name: format
          in: query
          description: json, csv, xlsx format download
          required: false
          type: string
          default: csv
        - name: from
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20170101'
        - name: query
          in: query
          description: query
          required: true
          type: string
        - name: sort
          in: query
          description: "\"entity.asc\", \"entity.desc\", \"date.asc\", \"date.desc\"; default: date.desc,entity.desc"
          required: false
          type: array
          items:
            type: string
          collectionFormat: multi
          default: 'date.desc,entity.desc'
        - name: to
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20190101'
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT:
            - global
      deprecated: false
  /news/entity/spss/download:
    get:
      tags:
        - news-controller
      summary: downloadEntitySPSS
      description: 검색어가 포함된 뉴스기사의 개체명 항목 CSV 다운로드(SPSS)
      operationId: downloadEntitySPSSUsingGET
      produces:
        - '*/*'
      parameters:
        - name: from
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20180101'
        - name: query
          in: query
          description: query
          required: true
          type: string
        - name: size
          in: query
          description: size
          required: false
          type: integer
          default: 1000
          format: int32
        - name: to
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20180211'
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT:
            - global
      deprecated: false
  /news/entity/summary:
    get:
      tags:
        - news-controller
      summary: getEntitySummary
      description: 검색어가 포함된 뉴스기사의 전체 개체명 항목 및 건수  
      operationId: getEntitySummaryUsingGET
      produces:
        - '*/*'
      parameters:
        - name: from
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20170101'
        - name: query
          in: query
          description: query
          required: true
          type: string
        - name: size
          in: query
          description: size
          required: false
          type: integer
          default: 100
          format: int32
        - name: to
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20190101'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/NewsNamedEntitySummaryList'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT:
            - global
      deprecated: false
  /news/related:
    get:
      tags:
        - news-controller
      summary: getRelatedTopic
      description: 검색어를 포함하는 전체 뉴스기사에서 입력한 키워드와 가장 연관성이 높은 관련 키워드
      operationId: getRelatedTopicUsingGET
      produces:
        - '*/*'
      parameters:
        - name: from
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20180101'
        - name: query
          in: query
          description: query
          required: true
          type: string
        - name: size
          in: query
          description: size
          required: false
          type: integer
          default: 100
          format: int32
        - name: to
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20180211'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Map«string,string»'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT:
            - global
      deprecated: false
  /news/related/download:
    get:
      tags:
        - news-controller
      summary: downloadRelatedTopic
      description: 검색어를 포함하는 전체 뉴스기사에서 입력한 키워드와 연관성이 높은 관련 키워드
      operationId: downloadRelatedTopicUsingGET
      produces:
        - '*/*'
      parameters:
        - name: bsize
          in: query
          description: 각 날짜별 가장 연관성이 높은 상위 개수
          required: false
          type: integer
          default: 10
          format: int32
        - name: from
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20170101'
        - name: query
          in: query
          description: query
          required: true
          type: string
        - name: size
          in: query
          description: 특정 하루 기준 가장 연관성이 높은 관련 키워드
          required: false
          type: integer
          default: 10000
          format: int32
        - name: to
          in: query
          description: \"yyyyMMdd\" format String
          required: false
          type: string
          x-example: '20181231'
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT:
            - global
      deprecated: false
  /news/related/rank/download:
    get:
      tags:
        - news-controller
      summary: downloadRelatedRank
      operationId: downloadRelatedRankUsingGET
      produces:
        - '*/*'
      parameters:
        - name: bsize
          in: query
          description: bsize
          required: false
          type: integer
          default: 10
          format: int32
        - name: format
          in: query
          description: format
          required: false
          type: string
          default: csv
        - name: from
          in: query
          description: String
          required: false
          type: string
          x-example: '20180101'
        - name: query
          in: query
          description: query
          required: true
          type: string
        - name: size
          in: query
          description: size
          required: false
          type: integer
          default: 10000
          format: int32
        - name: to
          in: query
          description: String
          required: false
          type: string
          x-example: '20180211'
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      security:
        - JWT:
            - global
      deprecated: false
...
securityDefinitions:
  JWT:
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  'Map«string,string»':
    type: object
    title: 'Map«string,string»'
    additionalProperties:
      type: string
...
  NewsNamedEntity:
    type: object
    properties:
      link:
        type: string
      named_entity_list:
        type: array
        items:
          type: string
      title:
        type: string
      written_time:
        type: string
    title: NewsNamedEntity
  NewsNamedEntityList:
    type: object
    properties:
      news_named_entity_list:
        type: array
        items:
          $ref: '#/definitions/NewsNamedEntity'
      total_hits:
        type: integer
        format: int64
    title: NewsNamedEntityList
  NewsNamedEntitySummary:
    type: object
    properties:
      doc_count:
        type: integer
        format: int64
      key_word:
        type: string
    title: NewsNamedEntitySummary
  NewsNamedEntitySummaryList:
    type: object
    properties:
      entity_name:
        type: string
      news_named_entity_summary_list:
        type: array
        items:
          $ref: '#/definitions/NewsNamedEntitySummary'
      num_total_terms:
        type: integer
        format: int64
    title: NewsNamedEntitySummaryList
...